[seatunnel-e2e] Add mongodb container to simplify spark-e2e and flink-e2e
