+++ seatunnel-e2e/pom.xml
@@ -34,6 +34,7 @@
         <module>seatunnel-flink-sql-e2e</module>
         <module>seatunnel-spark-connector-v2-e2e</module>
         <module>seatunnel-spark-e2e</module>
+        <module>seatunnel-containers</module>
     </modules>
 
     <properties>
