+++ seatunnel-e2e/seatunnel-flink-connector-v2-e2e/connector-mongodb-flink-e2e/src/test/resources/mongodb/mongodb_source_and_sink.conf
@@ -29,7 +29,7 @@ env {
 
 source {
   MongoDB {
-    uri = "mongodb://flink_e2e_mongodb:27017/test_db?retryWrites=true&writeConcern=majority"
+    uri = "mongodb://seatunnel_mongodb:27017/test_db?retryWrites=true&writeConcern=majority"
     database = "test_db"
     collection = "source_table"
     schema = {
@@ -59,7 +59,7 @@ transform {
 
 sink {
   MongoDB {
-    uri = "mongodb://flink_e2e_mongodb:27017/test_db?retryWrites=true&writeConcern=majority"
+    uri = "mongodb://seatunnel_mongodb:27017/test_db?retryWrites=true&writeConcern=majority"
     database = "test_db"
     collection = "sink_table"
   }
