+++ seatunnel-e2e/seatunnel-spark-connector-v2-e2e/connector-mongodb-spark-e2e/src/test/resources/mongodb/fake_to_mongodb.conf
@@ -60,7 +60,7 @@ transform {
 
 sink {
   MongoDB {
-    uri = "mongodb://spark_e2e_mongodb_sink:27017/test_db?retryWrites=true&writeConcern=majority"
+    uri = "mongodb://seatunnel_mongodb:27017/test_db?retryWrites=true&writeConcern=majority"
     database = "test_db"
     collection = "test_table"
   }
